modify the update colour functionality and slightly revise code

+++ client/components/containers/EditNote.js
@@ -1,7 +1,5 @@
 import React, { Component } from 'react'
-import { Button, Dropdown, Form, Grid, Input, Label, Modal, TextArea } from 'semantic-ui-react'
-
-import styles from '../../styles.js'
+import { Button, Dropdown, Form, Grid, Icon, Input, Label, Modal, TextArea } from 'semantic-ui-react'
 
 class EditNote extends Component {
   constructor(props){
@@ -37,13 +35,53 @@ class EditNote extends Component {
   }
 
   delete = () => {
-    this.props.deleteNote(this.state.note.id);
+    if (this.state.note.id){ // delete note if it has been created (no id if note wasn't added)
+      this.props.deleteNote(this.state.note.id);
+    }
     this.props.close();
   }
 
+  updateColour = (event) => {
+    let currentButton = event.currentTarget.id;
+    let selectedButton = document.querySelectorAll('.colour-selected');
+    if (selectedButton[0]){
+      selectedButton[0].classList.remove('colour-selected');
+    }
+    document.getElementById(currentButton).classList.toggle('colour-selected');
+
+    let elements = document.querySelectorAll('.' + this.state.note.colour);
+    for (let i = 0; i < elements.length; i++){
+      if (elements[i].classList.contains('modal')){
+        elements[i].classList.remove(this.state.note.colour);
+        elements[i].classList.add(currentButton);
+      }
+    }
+
+    if(this.state.note.colour !== currentButton){
+      this.setState({
+        note: {...this.state.note, ...{colour: currentButton}}
+      });
+    }
+  }
+
   render(){
+    const colours_1 = ['white', 'lightgreen', 'lightskyblue'];
+    const colours_2 = ['lightcoral', 'yellow', 'rosybrown'];
+
+    const colourList_1 = colours_1.map((colour, i) => {
+      return(
+        <Button key={i} circular icon='check' id={colour} className={this.state.note.colour === colour ? 'colour-selected' : ''} onClick={this.updateColour.bind(this)} />
+      )
+    })
+
+    const colourList_2 = colours_2.map((colour, i) => {
+      return(
+        <Button key={i} circular icon='check' id={colour} className={this.state.note.colour === colour ? 'colour-selected' : ''} onClick={this.updateColour.bind(this)} />
+      )
+    })
+
     return(
-      <Modal dimmer='inverted' open={this.props.open} onClose={this.close.bind(this)}>
+      <Modal dimmer='inverted' open={this.props.open} onClose={this.close.bind(this)} className={this.props.item.colour}>
         <Modal.Header>
           <Input id='title' fluid defaultValue={this.props.item.title} onChange={this.handleInputChange.bind(this)} />
         </Modal.Header>
@@ -64,9 +102,8 @@ class EditNote extends Component {
             <Grid.Column textAlign='left'>
               <Dropdown floating button className='icon inverted green bottom left' icon='paint brush' pointing>
                 <Dropdown.Menu>
-                  <Dropdown.Item icon='user' text='Details' />
-                  <Dropdown.Item icon='setting' text='Settings' />
-                  <Dropdown.Item icon='sign out' text='Logout' />
+                  <Dropdown.Item className='with-grid'>{colourList_1}</Dropdown.Item>
+                  <Dropdown.Item className='with-grid'>{colourList_2}</Dropdown.Item>
                 </Dropdown.Menu>
               </Dropdown>
               <Button inverted color='green' icon='add user' onClick={this.close.bind(this)} />
